+++ src/040-imports/__testdata/stdlib/lib/config.tact
